keep private classes out base file as this is most likely to be opened by end user
